refactor: converted focus from page to component
