+++ src/components/MyTimeComponents/Focus.tsx/focus.scss
